+++ src/main/java/co/edu/uniquindio/aplicacion/carro/CarroComprasService.java
@@ -1,5 +1,6 @@
 package co.edu.uniquindio.aplicacion.carro;
 
+import java.util.List;
 import java.util.Optional;
 
 import co.edu.uniquindio.dominio.carro.CarroCompras;
@@ -11,12 +12,14 @@ public class CarroComprasService {
     private CrearCarroComprasUseCase crearCarroCompras;
     private HallarCarroComprasUseCase hallarCarroCompras;
     private ValidarCarroComprasUseCase validarCarroCompras;
+    private HallarCarroComprasPorCiudadanoUseCase hallarCarroComprasPorCiudadano;
 
     public CarroComprasService(CarroComprasRepository repositorio) {
         // this.repositorio = repositorio;
         crearCarroCompras = new CrearCarroComprasUseCase(repositorio);
         hallarCarroCompras = new HallarCarroComprasUseCase(repositorio);
         validarCarroCompras = new ValidarCarroComprasUseCase(repositorio);
+        hallarCarroComprasPorCiudadano = new HallarCarroComprasPorCiudadanoUseCase(repositorio);
     }
 
     public CarroCompras craer(CarroCompras carroCompras) {
@@ -29,5 +32,9 @@ public class CarroComprasService {
 
     public Boolean validar(Integer id) {
         return validarCarroCompras.execute(id);
+    }
+
+    public List<CarroCompras> hallarPorCiudadano(Integer ciudadanoId) {
+        return hallarCarroComprasPorCiudadano.execute(ciudadanoId);
     }
 }