+++ src/main/java/co/edu/uniquindio/aplicacion/producto/ProductoService.java
@@ -1,5 +1,6 @@
 package co.edu.uniquindio.aplicacion.producto;
 
+import java.util.List;
 import java.util.Optional;
 
 import co.edu.uniquindio.dominio.producto.Producto;
@@ -10,11 +11,13 @@ public class ProductoService {
 
     private CrearProductoUseCase crearProducto;
     private HallarProductoUseCase hallarProducto;
+    private HallarProductosPorNombreUseCase hallarProductosPorNombre;
 
     public ProductoService(ProductoRepository repositorio) {
         // this.repositorio = repositorio;
         crearProducto = new CrearProductoUseCase(repositorio);
         hallarProducto = new HallarProductoUseCase(repositorio);
+        hallarProductosPorNombre = new HallarProductosPorNombreUseCase(repositorio);
     }
 
     public Producto crear(Producto producto) {
@@ -23,5 +26,9 @@ public class ProductoService {
 
     public Optional<Producto> hallarPorId(Integer id) {
         return hallarProducto.execute(id);
+    }
+
+    public List<Producto> hallarPorNombre(String nombre) {
+        return hallarProductosPorNombre.execute(nombre);
     }
 }