+++ src/main/java/co/edu/uniquindio/infraestructura/detalleproducto/memory/InMemoryDetalleProductoRespository.java
@@ -1,6 +1,7 @@
 package co.edu.uniquindio.infraestructura.detalleproducto.memory;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -35,6 +36,14 @@ public class InMemoryDetalleProductoRespository implements DetalleProductoReposi
     @Override
     public Optional<DetalleProducto> hallar(Integer id) {
         return Optional.ofNullable(datos.get(id));
+    }
+
+    @Override
+    public List<DetalleProducto> obtenerPorCarroCompras(Integer carroComprasId) {
+        return datos.values().stream()
+            .filter(dp -> dp.getCarroCompras().getId() == carroComprasId)
+            .toList()
+        ;
     }
 
     @Override